article_filter/requirements.txt
kafka-python
pymongo
dnspython
flask